refactor: :art: moves component constants into its own directory
