+++ frontend/src/components/BudgetLineItems/AllBudgetLinesTable/AllBudgetLinesTable.constants.js
@@ -1,0 +1,3 @@
+export const All_BUDGET_LINES_TABLE_HEADINGS = ["Description", "Agreement", "Need By", "FY", "CAN", "Total", "Status"];
+
+export const BLIS_PER_PAGE = 10;
