+++ frontend/src/components/BudgetLineItems/AllBudgetLinesTable/AllBudgetLinesTable.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import _ from "lodash";
 import Table from "../../UI/Table";
 import AllBLIRow from "./AllBLIRow";
-import { All_BUDGET_LINES_TABLE_HEADINGS, BLIS_PER_PAGE } from "../../../constants";
+import { All_BUDGET_LINES_TABLE_HEADINGS, BLIS_PER_PAGE } from "./AllBudgetLinesTable.constants";
 import PaginationNav from "../../UI/PaginationNav/PaginationNav";
 
 /**
