+++ frontend/src/components/BudgetLineItems/BudgetLinesTable/BudgetLinesTable.constants.js
@@ -1,0 +1,1 @@
+export const BUDGET_LINE_TABLE_HEADERS = ["Description", "Need By", "FY", "CAN", "Amount", "Fee", "Total", "Status"];
