+++ frontend/src/components/BudgetLineItems/BudgetLinesTable/BudgetLinesTable.jsx
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import TotalSummaryCard from "../TotalSummaryCard/TotalSummaryCard";
 import Table from "../../UI/Table";
 import "./BudgetLinesTable.scss";
-import { BUDGET_LINE_TABLE_HEADERS } from "../../../constants";
+import { BUDGET_LINE_TABLE_HEADERS } from "./BudgetLinesTable.constants";
 import BLIRow from "./BLIRow";
 
 /**
